refactor: Use Jetpack Navigation Component
with Safe Args gradle plugin for type safety in both navigation and arguments for navigation destinations.

This is quite a refactor. Previously, MainActivity had to manually use the FragmentManager so we had to use a callback interface to get it to navigate after a list item was clicked in the CrimeListFragment.

Now, MainActivity simply hosts the NavHostFragment that points to the navGraph, where all the action is defined. For example, what the starting destination should be.

CrimeListFragment knows how to navigate to a CrimeDetailFragment when a crime list item is clicked: onCrimeItemClicked().

#15