feature: Let user pick date with a DatePicker
in CrimeDetailFragment when user clicks on the date button.

This is the vanilla solution from the book. It uses several methods that are now marked as deprecated: the ones for targetFragment use.

This solution is quite elaborate. It uses again a Callbacks interface that the starting Fragment must implement so that the DialogFragment can call this callback to send the date back to the CrimeDetailFragment.

Also, this solution uses a manual arguments bundle to bundle the current date of the crime to the DatePickerDialog as a starting date. It does not use the Jetpack Navigation Component, nor its SafeArgs.

#15