refactor: Use FragmentResult to transfer selected date to CrimeDetailFragment
The method of setTargetFragment and getTargetFragment are deprecated. We use the recommended way:

We use FragmentManager.setFragmentResult(request_key, bundle) in DatePickerFragment that automatically calls all listeners that were set by FragmentManager.setFragmentResultListener(request_key, ...)

It's nice to have a built-in lifecycle-aware observer pattern implementation in the FragmentManager itself to transfer data back to caller fragments.

Now we don't have to use the Callbacks interface any more :)

#15