refactor: clearer distinction between responsibilities of CrimeDetailFragment and DatePickerFragment
The CrimeDetailFragment knows 2 things about DatePickerFragment:
a) how to create and start a new instance by saying: "here, you have an initial date to display, and a request key that you should use for my fragmentResult", and
b) that to get the date from the result bundle, you have to call the companion object (static) method getDateFrom(Bundle)

The rest is the concern of DatePickerFragment:
It receives the initial date and the request key, packs them in an arguments bundle for itself. It displays the initial date and when the user changes the date, it uses the request key to set the right fragment result in the parentFragmentManager.
And since it has bundled the result itself, it also knows how to un-bundle the date so it offers that knowledge as an API call to getDateFrom(Bundle)

#15