Implement iterator pattern
Pancake House and Diner do agree on the MenuItem but don't agree on their implementations of their respective menus and both have invested too much into their implementations to change either of them.

When the waitress wants to list the items of both menus, she must know the concrete implementations of both. At least without the iterator pattern.

Both diners agree on an Iterator interface and both implement methods that return an implementation.
Now the waitress only needs to know about Iterator and uses it. She doesn't care how this interface is implemented, she only cares that it works.

#16