feat: Add a third Menu: CafeMenu
Now Waitress prints out 3 Menus using the Iterator coming from the Menu Interface method. All 3 Menus are implemented differently, but Waitress neither knows that nor does she care.

But Waitress' implementation violates the Open-Closed-Principle. When we added the third Menu, we had to modify the printMenu() method and we'll have to do this every time a new Menu is added.

#16