fix: update view when data changes
We implemented the sound view model of MVVM manually. And the Recyclerview ViewHolder (SoundHolder) now holds a binding object instead of a view and this binding object is associated with its own view model object.

During an onBindViewHolder(), we simply associate a new Sound model object with the view model. Now we're using BaseObservable and Bindable to make the data binding observe the changes and display the correct and current data.

#17