feat: API Calls now possible
We set up Retrofit2 as the REST API library. FlickrFetcher works as a Repository: it's an abstraction layer between our code and the API requests.

Also, the API key is stored in `local.properties` as `FLICKR_API_KEY="XXXXXXXXXX...XXX"` and inserted into the project during build. This way we don't have to commit and upload the API key to Github. #20

#19