feat: Download and display actual pictures
We added code that downloads the pictures and updates the UI (recyclerview items), while handling the main/UI thread vs the downloading background thread.

This code is a mess! I followed the tutorial because it would take me a long time refactoring this into something more modern and easier to understand. And I guess, this is my implicit homework :).

(This code contains lots of deprecated stuff and things that give me stomach pains as a software developer.)

I'm thinking of a nice ViewModel with LiveData and coroutines for background downloading and image processing.

#19