refactor: Encapsulate state in State Implementations
The previous solution had several problems. One of the most visible is the collection of huge if-else blocks. They are a nightmare to maintain and extend. This solution is so poor, it's not even OOP.

The new implementatino defines a State interface that is implemented by the concrete states. The main idea of a state is that it encapsulates behavior: when a certain state is active, certain behaviors are possible (or not).

The different states are mostly independent of each other, so when the behavior of a single state changes, the others are not touched by it. Also, we can add new states by defining new State implementations.

#22