BroadcastReceiver POC
Make PollWorker send out a Broadcast Intent that is intended for the new NotificationReceiver. It is a standalone receiver registered in AndroidManifest.xml and executed when it receives this broadcast intent.

This is the first step in making sure, the notification does *not* appear when the user is in the app.

#19