refactor: several components
Using VisibleFragment as intermediary Class for PhotoGalleryFragment was confusing, but it did encapsulate the management of the dynamic notification broadcast receiver. Now it's in a separate class that is a LifecycleObserver and observes the PhotoGalleryFragment and manages the registration of the receiver accordingly.

Make NotificationReceiver more independent. It now knows nothing about the other app parts. It simply extracts a notification from the broadcast intent and shows it via the notification manager. It knows how to get the broadcast and requestcode from the intent, so other components wanting to show a notification need to refer to this class on how to put them in the intent.

#19