refactor: PollWorker and NotificationReceiver
NotificationReceiver now knows how to create and pack intents that will be sent to it later via broadcast. It already knows how to unpack it. Now that knowledge is encapsulated here so that no other component needs to know how to pack and unpack the reuest code and notification in the broadcast intent.

PollWorker: renamings and minor refactorings

#19