feat: View photo webpage in-app
Set up WebView with a default WebViewClient so that it displays the webpage in the app's fragment.
(Without a WebViewClient, WebView would send out a view intent for the browser when the user clicks on a link or the webpage gets redirected, e.g. to a mobile site version.)

This version still needs some tweaking. For example, the webpage is not loaded in mobile form and does not fit on the screen horizontally.

#19