Set up tests for TasksViewModel
TasksViewModel depends on DefaultTasksRepository. So we need to inject that dependency in order to inject a test double in tests.
We extract a TasksRepository interface from DefaultTasksRepository and create FakeTasksRepository that implements that interface and is now a viable double.

We now inject DefaultTasksRepository into the constructor of TasksViewModel. But now we need a custom ViewModelFactory because ViewModelProvider does not know how to instantiate viewmodels without no-args constructors.

We do the same steps for TaskDetailViewModel so that we can test it later.

#25