Does a coroutine gets cancelled when I don't want it?
There is an example with a pure kotlin project where the main function starts a coroutine and this coroutine is not finshed because it is canceled. It is canceled because the main function ends and finishes also the corresponding coroutine scopes.

In Android this is different, slightly. We do use coroutine scopes that are far greater than a single function. Like lifecycleScope. Those scopes do have an end of their own "lifecycle" and then all child coroutines do get canceled.

#30