Implement Service Locator Pattern
Fragments and Activities are created by Android without influence of us, the developers. So we can't simply add something like TasksRepository as a manual dependency injection in the constructor.
Instead, we have to inject that dependency in another way. Here we implement the Service Locator Pattern with a singleton (object) that TodoApplication uses to instantiate a TasksRepository with the Application's context. Now the TodoApplication offers TasksRepository as an instance variable.

We will create a double for the ServiceLocator in the tests that will then provide doubles for TasksRepository.

#25