Add DatabaseVideo as a database entity
This app teaches how to separate the models:
- Domain model is what the app business logic uses exclusively
- Data transfer model is what the network package uses to receive/parse the network requests
- and now the Database model are the entities for Room

It makes a lot of sense to separate them into separate models because all of them have very distinct responsibilities although they seem to look similar.
If we put them in a single object, it would be cluttered with the network transfer and database stuff.

#34