Add VideosRepository with clear API
VideosRepository is responsible to provide Videos to the business logic, that doesn't care where the data comes from.
So we can hugely simplify the code of DevByteViewModel by simply toggling a refresh when the viewModel is created and delegating the rest to the Repository.

The API of this repository is super simple:
A) it exposes a LiveData<List<Video>> property that the viewModel can use directly to expose to the view for observation.
B) it has a suspend function refreshVideos() and requires the client code to use it to manually toggle a refresh.

This repository uses the database as the single source of truth. Meaning that it provides only data from the database. On a refresh, we simply update the data on the database, and Room takes care of propagating that change to the LiveData observers for us. :)

#34