Refactor mars property detail screen
Binding adapters are great if they are self-contained. But we want to use string resources to format the data for the view.
So we shift the display format logic from binding adapters to a view model that has a reference to application and thus can getString().

We display the data slightly differently depending if it's for rent or for sale.

The solution uses LiveData in DetailViewModel. But it's not needed, since in this example the selected mars property data never changes. (If it were fetched by Room from a database, LiveData would be the way to go)

#35