Add navigation arguments for GameWonFragment
We want to have the number of questions and the number of correct questions in GameWonFragment so that we can either display them, or share them with others.

SafeArgs is a perfect solution for that.

SafeArgs is a gradle plugin that generates the correct action and directions classes.

Using their navigation actions, we are required to give arguments that have no default value. This is enforced by the compiler :)

This version of GameWonFragment reads the safeArgs with
  val args = GameWonFragmentArgs.fromBundle(arguments!!)

(but there is also the by navArgs() extension delegate)

#37