Add a navigation drawer to the app
A navigation drawer can be activated by swiping from outside a vertical screen side, or when in the home destination clicking the "up" button that is displayed with a hamburger icon.

NavigationUI helps us connect the components together: action bar, drawer layout (parent layout for content and drawer), the navigation drawer, and the NavController.

#37