Refactor refreshTitle: use coroutines for blocking calls
We changed the old implementation with "manual" background thread use and a generous use of callbacks with coroutines that make the code a lot easier because it now looks and feels sequential and allows to use known constructs like a try-catch block.

This version still uses thread-blocking calls for network and database, but since it's done on Dispatchers.IO, only a IO thread gets blocked and the parent suspend function (probably on Main thread) gets suspended (as expected) in the meantime.

Since both network and database calls are old-fashioned thread blocking, they won't notice if the job of the coroutine has been cancelled and will execute fully. To mitigate that problem, we throw in a suspending yield() between them that allows other coroutines to be executed and also checks if the connected job is still active.

#38