Use runBlockingTest() and InstantTaskExecutorRule
runBlockingTest is similar to runBlocking and executes the coroutines in a blocking fashion effectively making it single threaded and more deterministic. The test flavor of those offers a virtual clock that is advanced immediately when it encounters a delay() which allows tests to run faster without having to actually wait. It also allows me to manipulate this virtual clock with advanceTimeBy.

We also use InstantTaskExecutorRule that swaps out the custom background threading of Room and Retrofit for something more deterministic.

#38