Implement custom sort order
We want to display certain plants at the beginning of the list in recyclerview. For this, we use the liveData {} extension function backed by coroutines where we fetch a list defining the custom order, apply it to the list and emit the result
We do the same with the list with growth zone.

#39