Setup parallel data fetch mechanism using Flow
Room supports the use of Flow to fetch data from the database, we implement that in parallel to the standard liveData way.
But in the view model, we convert the flow to a LiveData so that the UI can observe the LiveData without knowing that Flow is used in the background.

#39