Add sort order for plant list using flow
we use the map{} extension function of flow to call two suspending functions, a) get the sort Order from network (or cache), and b) apply this sort in a main-safe suspend function.

This way is in contrast to the plantsFlow way of using .combine() where the network request and database lookup are done concurrently. Here, they are run inside a coroutine sequentially (but suspending), which makes the code more clear.

It depends on the context and what you want to decide which to use.

#39