Move update cache logic from onClickListeners to flow
Whenever growZone changes, we want to fetch the right plants for the new zone. LiveData is only partly convenient for that. In particular, it requires a view to observe it, otherwise it won't emit anything. This means you can't observe it from the ViewModel because a view model doesn't have the required viewLifecycleOwner (and it would be a bad idea because of separations of concern).
Enter Flow. Flow offers the same benefits of LiveData in that it automatically emits data when updated, and you can observe that change from any data layer like viewModel or Repository.

#39