Refactor GameViewModel nextWord()
so far, it did several clashing things, it updated the word in the view model, checked if the same should still run and returned that boolean.

It clearly violated the principle of either tell or query a function.

Now it updates the view model status according to the rules and returns nothing. But it also sets a visible boolean variable telling the fragment if the game is still on. Now the fragment can react accordingly to this new variable.

#41