Tweak & Refactor GameOverDialogFragment
the viewModels extension function can take an `ownerProducer` that defines a ViewModelStoreOwner. In GameFragment we can keep `by viewModels()` and keep the scope of the view model tied to this fragment as it should. In the Dialog fragment we can tell `by viewModels(..)` to use the parent fragment as the ViewModelStoreOwner and use its scope. :)

#41