Setup navigation and app bar titles
There were only actions missing in the navigation graph. Use those actions with navController to execute the navigation.

Also we have to inconveniently retrieve the navController in onCreate of the activity in order to connect it with the app bar (action bar) so that the bar will display labels set for the destinations in the nav graph.

#45