Introduce shared view model
OrderViewModel: for the fragments of this ordering process. Add it as data binding data in the layout files and drive the view from this view model.

It is a bit elaborate to set up the data binding that doesn't assume any (starting) state but always checks for the data. But that is the principle of driving UI from a (view-)Model and with that we get full UI restoration on configuration changes and when navigating the fragment backstack.

#45