Add price update during ordering process
The view model now calculates the price from the other data, and we use data binding to bind this data in the layouts.

The data binding in the layout correctly reads the value during creation, but for continuous observation of the data binding, we need to give the binding object a lifecycleOwner in the form of viewLifecycleOwner of the fragment. Data binding (any lifecycle-aware observation) only triggers for the lifecyle owner being in a right state.

#45