Refactor: use data binding to set button click listeners
The button click listeners have to be defined in the fragments instead of the view model.

A fragment can easily set its own click listener, but using data binding is even more elegant. For this we add the corresponding fragment as a data binding variable in the layouts, in addittion to the (shared) viewModel.

Now we can easily set the onClick to a method from the fragments and make the fragment code clean.

#45