Fix back button on start screen
Navigating with the navController puts a new instance of the fragment that we navigate to, to the back stack. So we have to include the original instance of the starting fragment in the popped fragments so that afterwards the navigation can add a new instance that is the only one.

Otherwise the app won't close on back button press, but instead "navigate" to the previous starting fragment.

#45